agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
langgraph>=0.2.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
